src/ingest/requirements.txt
httpx>=0.27,<0.28
google-auth>=2,<3
pypdf>=4,<5
